+++ ContentPatcher/release-notes.md
@@ -5,6 +5,7 @@
 * Added new tokens:
   * `HasReadLetter`: whether the player has opened a given mail letter.
   * `HasValue`: whether a token has a non-blank value, like `HasValue:{{spouse}}`.
+  * `IsCommunityCenterComplete`: whether all bundles in the community center are completed.
   * `IsMainPlayer`: whether the player is the main player.
 * Tokens can now be nested (like `Hearts:{{spouse}}`).
 * Tokens can now be used in `Fields` keys.